Fixes#53
Arrow to Top placed  in projects component.

Appears after 2500px down.

Some bugs fixed
